chore(albwaf): rebase changes
Relates to STACKITCLI-443
