+++ internal/cmd/beta/alb-waf/alb-waf.go
@@ -1,6 +1,7 @@
 package albwaf
 
 import (
+	customrulegroup "github.com/stackitcloud/stackit-cli/internal/cmd/beta/alb-waf/custom-rule-group"
 	managedruleset "github.com/stackitcloud/stackit-cli/internal/cmd/beta/alb-waf/managed-rule-set"
 	"github.com/stackitcloud/stackit-cli/internal/pkg/args"
 	"github.com/stackitcloud/stackit-cli/internal/pkg/types"
@@ -25,5 +26,6 @@ func NewCmd(params *types.CmdParams) *cobra.Command {
 func addSubcommands(cmd *cobra.Command, params *types.CmdParams) {
 	cmd.AddCommand(
 		managedruleset.NewCmd(params),
+		customrulegroup.NewCmd(params),
 	)
 }
