+++ internal/cmd/beta/alb-waf/custom-rule-group/generate-payload/generate_payload.go
